+++ packages/moonstone/CHANGELOG.md
@@ -17,6 +17,7 @@ The following is a curated list of changes in the Enact moonstone module, newest
 
 ### Fixed
 
+- `moonstone/Marquee` to re-evaluate RTL orientation when its content changes
 - `moonstone/VirtualList` to restore focus on short lists
 - `moonstone/ExpandableInput` to expand the width of its contained `moonstone/Input`
 - `moonstone/InputDecorator` support for `dismissOnEnter`